requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
attrs==25.3.0
beautifulsoup4==4.13.4
black==25.1.0
blinker==1.9.0
browser-cookie3==0.20.1
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
click-default-group==1.2.4
condense-json==0.1.3
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
gemini-webapi==1.14.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.73.0
grpcio-status==1.71.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
ijson==3.4.0
isort==6.0.1
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
llm==0.26
llm-gemini==0.22
loguru==0.7.3
lz4==4.4.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mypy_extensions==1.1.0
narwhals==1.42.1
numpy==2.3.0
openai==1.86.0
orjson==3.10.18
packaging==24.2
pandas==2.3.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
proto-plus==1.26.1
protobuf==5.29.5
puremagic==1.29
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycryptodomex==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
python-ulid==3.0.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rich==14.1.0
rich-toolkit==0.14.9
rignore==0.6.4
rpds-py==0.25.1
rsa==4.9.1
sentry-sdk==2.34.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
sqlite-fts4==1.0.3
sqlite-migrate==0.1b0
sqlite-utils==3.38
starlette==0.47.2
streamlit==1.45.1
tabulate==0.9.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
